% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare_detects}
\alias{compare_detects}
\title{Compare two Detection Dataframes}
\usage{
compare_detects(a1, a2)
}
\arguments{
\item{a1}{unfiltered detection dataframe}

\item{a2}{filtered detection dataframe}
}
\value{
A dataframe of detections from the original dataframe missing from
the filtered dataframe. Spurious detections.
}
\description{
This function compares two detection dataframes from filtered acoustic
telemetry detections and identifies the filtered out detections. The first
input is the original dataframe, the second is the filtered dataframe. Then
it creates a dataframe of detections which are missing from the filtered
dataframe.
}
\examples{
# Compare detections before and after filtering
compare_detects(raw_ats, filter_fish_detects)
}
