% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{fish_detects}
\alias{fish_detects}
\title{Example fish detection data}
\format{
\subsection{\code{fish_detects}}{

A dataframe with 10,000 rows and 3 columns
\describe{
\item{ReceiverSN}{the serial number of the acoustic receiver which heard
the detection}
\item{DateTime_Local}{The local date and time of the detection
tz = "America/Los_Angeles"}
\item{Tag_Code}{The hexadecimal value of the acoustic tag ID code}
}
}
}
\usage{
fish_detects
}
\description{
Example tag detection data from CalFishTrack representing fish released in
2021.
}
\keyword{datasets}
