% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fish_fields.R
\name{get_fish_fields}
\alias{get_fish_fields}
\title{Get A List of Fish Related Fields from ERDAPP}
\usage{
get_fish_fields()
}
\value{
A vector of potential fish fields which the user may review to
retrieve specific field indices
}
\description{
This function searches the California Fish Tracking ERDAPP Database to create
a list of all potential fields related to fish metrics. Used to identify
important fish data fields to add to detection data in the add_fish function.
}
\examples{
fish_fields <- get_fish_fields
}
