% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rcvr_fields.R
\name{get_rcvr_fields}
\alias{get_rcvr_fields}
\title{Get A List of Receiver Related Fields from ERDAPP}
\usage{
get_rcvr_fields()
}
\value{
A vector of potential receiver metadata fields which the user may review
}
\description{
This function searches the California Fish Tracking ERDAPP Database to create
a list of all potential fields related to acoustic receiver metadata.
Used to identify important metadata fields to include when adding receiver
data in the join_rcvr_data function.
}
\examples{
# View a list of available receiver fields
get_rcvr_fields()
}
