% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref_tags.R
\name{get_reference_tags}
\alias{get_reference_tags}
\title{Get A List of Reference (Beacon) Tags from ERDAPP}
\usage{
get_reference_tags()
}
\value{
A vector of beacon tags hexadecimal IDs
}
\description{
This function searches the California Fish Tracking ERDAPP Database to create
a list of beacon tag hexadecimal IDs. Used in the prefilter to separate
beacon tags from tagged fish.
}
