% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{rcvr_fields}
\alias{rcvr_fields}
\title{A List of Important Receiver Related Fields from ERDAPP}
\format{
A vector of potential receiver metadata fields which the user may review
}
\usage{
rcvr_fields
}
\description{
A vector of acoustic receiver metadata fields which are needed, at a minimum,
to run the join_rcvr_data function
}
\keyword{datasets}
