% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lotek_v2.R
\name{read_lotek_v2}
\alias{read_lotek_v2}
\title{Read a Lotek Acoustic Receiver File Version 2}
\usage{
read_lotek_v2(path, file, sensor = FALSE, timezone = "America/Los_Angeles")
}
\arguments{
\item{path}{the path to the folder containing the desired file}

\item{file}{the path of the desired file}

\item{sensor}{Logical, does the tag have an environmental sensor? default is FALSE}

\item{timezone}{the Olsen Named time zone, default is "America/Los_Angeles"}
}
\value{
A dataframe converting the raw detection data into rows of detections
}
\description{
This function takes a V2 raw acoustic detection file generated by a Lotek
JSATS receiver and processes it into a dataframe which can be used by
the filtering functions in this package.This is called within read_jsats().
}
\examples{
# see read_ats or read_tekno for example usage

}
