% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{register_parallel}
\alias{register_parallel}
\title{Register Parallel}
\usage{
register_parallel(files)
}
\arguments{
\item{files}{unfiltered detection dataframe}
}
\value{
Your machine is now ready to use parallel processing for this package
}
\description{
This function registers the various necessary packages on your machine for
for the parallel processing of files.
}
