% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4h_filter_all.R
\name{second_filter_4h}
\alias{second_filter_4h}
\title{Four Hit Filter for Any Detection File}
\usage{
second_filter_4h(fish_file)
}
\arguments{
\item{fish_file}{a dataframe of detections retrieved from add_fish()}
}
\value{
A dataframe which has been filtered to remove false positives
}
\description{
This function takes any acoustic receiver detection dataframe generated from
the add_fish() function and filters it a second time to remove any remaining
multipath detections. Then applies an algorithm to assess and filter all
remaining detections based on a four hit filter.
}
