% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_3.R
\name{second_filter_tekno}
\alias{second_filter_tekno}
\title{Two Hit Filter for Teknologic Detections}
\usage{
second_filter_tekno(fish_file)
}
\arguments{
\item{fish_file}{a dataframe of detections retrieved from add_fish()}
}
\value{
A dataframe which has been filtered to remove false positives
}
\description{
This function takes a Teknologic detection dataframe generated from the
add_fish() function and filters it a second time to remove any remaining
multipath detections, and then check the remaining detections by comparing
the time between each detection to ensure it is less 4x the stated pulse rate
interval. It additionally checks that all detections have a frequency
between 390 and 445 kHz and that the frequency of all detections are within
55kHz of each other. Called by second_filter().
}
