% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4h_second_filter_tekno.R
\name{second_filter_tekno_4h}
\alias{second_filter_tekno_4h}
\title{Four Hit Filter for Tekno Detections}
\usage{
second_filter_tekno_4h(fish_file)
}
\arguments{
\item{fish_file}{a dataframe of detections retrieved from add_fish()}
}
\value{
A dataframe which has been filtered to remove false positives
}
\description{
This function takes an Tekno detection dataframe generated from the
add_fish() function and filters it a second time to remove any remaining
multipath detections, and then check the remaining detections by comparing the time
between detections, for a rolling window of 4 detections to ensure it is less
16.6x the stated pulse rate interval. It additionally checks that all detections have a frequency
between 390 and 445 kHz and that the frequency of all detections are within
55kHz of each other. Called by second_filter_4h().
}
\examples{
# Apply a four-hit Teknologics filter to a prefiltered dataset with fish and tag
# attributes
second_filter_tekno_4h(filter_fish_detects)
# No detections are valid
}
