% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_2.R
\name{add_fish}
\alias{add_fish}
\title{Add Fish Data from ERDAPP to a Detection Dataframe}
\usage{
add_fish(prefilter_file, fish)
}
\arguments{
\item{prefilter_file}{a prefiltered detection dataframe from prefilter()}

\item{fish}{a dataframe of fish data rectrieved from get_fish_data()}
}
\value{
A filtered dataframe converting the raw detection data into rows of detections
}
\description{
This function takes a prefiltered detection dataframe from prefilter() and
joins it to fish data downloaded from ERDAPP using the get_fish_data()
function. Detections are then filtered durther based on the date and time of
tag release and expected battery life. Detections occuring before release of
the tag or after 2x the expected battery life are removed.
}
