% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_receiver.R
\name{join_rcvr_data}
\alias{join_rcvr_data}
\title{Add Receiver Data to a Completely Filtered Detection File}
\usage{
join_rcvr_data(final_file, rcvr_data)
}
\arguments{
\item{final_file}{a dataframe of detections retrieved from second_filter}

\item{rcvr_data}{a dataframe of receiver metadata retrieved from get_rcvr_data}
}
\value{
A dataframe to which receiver metadata has been added.
}
\description{
This function takes any acoustic receiver detection dataframe generated from
the second_filter() function and joins it with receiver metadata generated by
the get_rcvr_data() function.
}
