% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2h4h_second_filter.R
\name{second_filter_2h4h}
\alias{second_filter_2h4h}
\title{Two or Four Hit Filter for Any Detection File}
\usage{
second_filter_2h4h(fish_file)
}
\arguments{
\item{fish_file}{a dataframe of detections retrieved from add_fish()}
}
\value{
A dataframe which has been filtered to remove false positives
}
\description{
This function takes any acoustic receiver detection dataframe generated from
the add_fish() function and filters it a second time to remove any remaining
multipath detections. Then depending on the type of receiver file applies an
algorithm to assess and filter all remaining detections based on 2 or 4 hits.
}
\examples{
# Apply a 2 or 4 hit filter based on the technology
second_filter_2h4h(filter_fish_detects)
# No detections are valid, 15 tags accounted for a total of 30 spurious detects

}
